requirements.txt
streamlit
pillow
numpy
torch
pandas
torchvision
xgboost
git+https://github.com/PyTorchLightning/pytorch-lightning.git